requirements.txt
streamlit==1.32.0
beautifulsoup4==4.12.3
requests==2.31.0
faiss-cpu==1.7.4
openai>=1.12.0
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.1
tiktoken==0.6.0
urllib3==2.2.1
tqdm==4.66.2
httpx>=0.24.1 